# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSummary {
  total: number;
  active: number;
  gifts: number;
  testers: number;
  subscriptions: number;
  lifetime: number;
  recentIssuances: number;
}

export interface Licence {
  id: string;
  email: string;
  plan: string;
  source: string;
  active: boolean;
  /** @nullable */
  paidThrough?: string | null;
  purchasedAt: string;
  deviceLimit: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  stripeCheckoutSessionId?: string | null;
  deviceCount: number;
}

export interface LicenceInput {
  email: string;
  source: string;
  plan: string;
  /** @nullable */
  paidThrough?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface LicenceCreationResponse {
  licence: Licence;
  licenceKey: string;
}

export interface Device {
  id: string;
  hardwareId: string;
  model: string;
  osVersion: string;
  firstSeenAt: string;
  lastSeenAt: string;
  active: boolean;
}

export interface AuditRow {
  id: string;
  action: string;
  createdAt: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  details?: string | null;
}

export type LicenceDetail = Licence & {
  devices: Device[];
  auditRows: AuditRow[];
};

export interface LicenceStatusUpdate {
  active: boolean;
}

export type GetAdminLicencesParams = {
search?: string;
};

export type DeleteAdminLicenceDevice200 = {
  success?: boolean;
};