# lib/api-zod/src/generated/types/adminSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSummary {
  total: number;
  active: number;
  gifts: number;
  testers: number;
  subscriptions: number;
  lifetime: number;
  recentIssuances: number;
}