# lib/api-zod/src/generated/types/auditRow.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditRow {
  id: string;
  action: string;
  createdAt: string;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  details?: string | null;
}