# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  id: string;
  hardwareId: string;
  model: string;
  osVersion: string;
  firstSeenAt: string;
  lastSeenAt: string;
  active: boolean;
}