# lib/api-zod/src/generated/types/licence.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Licence {
  id: string;
  email: string;
  plan: string;
  source: string;
  active: boolean;
  /** @nullable */
  paidThrough?: string | null;
  purchasedAt: string;
  deviceLimit: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  /** @nullable */
  stripeCheckoutSessionId?: string | null;
  deviceCount: number;
}