# lib/api-zod/src/generated/types/licenceCreationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Licence } from './licence';

export interface LicenceCreationResponse {
  licence: Licence;
  licenceKey: string;
}