# lib/api-zod/src/generated/types/licenceDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditRow } from './auditRow';
import type { Device } from './device';
import type { Licence } from './licence';

export type LicenceDetail = Licence & {
  devices: Device[];
  auditRows: AuditRow[];
};