# lib/api-zod/src/generated/types/licenceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LicenceInput {
  email: string;
  source: string;
  plan: string;
  /** @nullable */
  paidThrough?: string | null;
  /** @nullable */
  note?: string | null;
}